PM / Hibernate: Include storage keys in hibernation image on s390
For s390 there is one additional byte associated with each page,
the storage key. This byte contains the referenced and changed
bits and needs to be included into the hibernation image.
If the storage keys are not restored to their previous state all
original pages would appear to be dirty. This can cause
inconsistencies e.g. with read-only filesystems.

Signed-off-by: Martin Schwidefsky <schwidefsky@de.ibm.com>
Signed-off-by: Rafael J. Wysocki <rjw@sisk.pl>

Conflicts:
	arch/s390/Kconfig
	kernel/power/Kconfig

Change-Id: I2081070877e9ec08088128c4955c600414895d96
Signed-off-by: Samuel Pascua <pascua.samuel.14@gmail.com>